add tests array, map, record, full name; fix recursive bug

+++ BaijiCore/BaijiCore/Schema/BJUnionSchema.m
@@ -47,7 +47,7 @@
     return self;
 }
 
-- (int)count {
+- (NSUInteger)count {
     return [self.schemas count];
 }
